Change type of all variables: from int to float

+++ src/Ability/RogueAbility.cpp
@@ -8,14 +8,14 @@ void RogueAbility::attack(Unit* attacker, Unit* enemy) {
     }
 
     enemy->ensureIsAlive();
-    int newEnemyHitPoint = (enemy->getState())->getHitPoints() - (attacker->getState())->getDamage();
+    float newEnemyHitPoint = (enemy->getState())->getHitPoints() - (attacker->getState())->getDamage();
 
     (enemy->getState())->takeDamage(newEnemyHitPoint);
 }
 
 void RogueAbility::counterAttack(Unit* counterAttacker, Unit* enemy) {
     std::cout << "RogueAbility COUNTER_ATTACK" << std::endl;
-    int newHitPoints = (enemy->getState())->getHitPoints() - (counterAttacker->getState())->getDamage()/2;
+    float newHitPoints = (enemy->getState())->getHitPoints() - (counterAttacker->getState())->getDamage()/2;
 
     (enemy->getState())->takeDamage(newHitPoints);
 }
