+++ src/Ability/VampireAbility.cpp
@@ -9,7 +9,7 @@ void VampireAbility::attack(AbstractVampire* attacker, Unit* enemy) {
     }
 
     enemy->ensureIsAlive();
-    int newEnemyHitPoint = (enemy->getState())->getHitPoints() - (attacker->getState())->getDamage();
+    float newEnemyHitPoint = (enemy->getState())->getHitPoints() - (attacker->getState())->getDamage();
 
     if ( newEnemyHitPoint < 0 ) {
         (attacker->getState())->setHitPoints((enemy->getState())->getHitPoints());
@@ -43,7 +43,7 @@ void VampireAbility::attack(Unit* attacker, Unit* enemy) {
 
 void VampireAbility::counterAttack(Unit* counterAttacker, Unit* enemy) {
     std::cout << "VampireAbility Vampire_counterATTACK" << std::endl;
-    int newHitPoints = (enemy->getState())->getHitPoints() - (counterAttacker->getState())->getDamage()/2;
+    float newHitPoints = (enemy->getState())->getHitPoints() - (counterAttacker->getState())->getDamage()/2;
 
 std::cout << newHitPoints << std::endl;
     if ( newHitPoints < 0 ) {
